a strange MC implementation.
we actually output a point cloud.  but, mostly I refer to how the
computation is done: instead of lookups in some giant table, we
just examine all edges to see if the surface falls between them,
and output a point if so.

valgrind is complaining about sample(x,y,z), saying it gives an
'invalid read'.  if i convert it to a function which does the same
thing, it's magically okay.  not sure what's going on, yet ...